--- lib/minitest-metadata.rb
@@ -1,2 +1,0 @@
-require "minitest-metadata/filters"
-require "minitest-metadata/patch"
